+++ src/Unity.Mathematics/math.cs
@@ -4062,6 +4062,214 @@ namespace Unity.Mathematics
         public static double csum(double4 x) { return (x.x + x.y) + (x.z + x.w); }
 
         /// <summary>
+        /// Computes the square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static float square(float x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static float2 square(float2 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static float3 square(float3 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static float4 square(float4 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static double square(double x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static double2 square(double2 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static double3 square(double3 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static double4 square(double4 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x)</c> is positive. For example, <c>square(46341)</c>
+        /// will return <c>-2147479015</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int square(int x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x)</c> is positive. For example, <c>square(new int2(46341))</c>
+        /// will return <c>new int2(-2147479015)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int2 square(int2 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x)</c> is positive. For example, <c>square(new int3(46341))</c>
+        /// will return <c>new int3(-2147479015)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int3 square(int3 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x)</c> is positive. For example, <c>square(new int4(46341))</c>
+        /// will return <c>new int4(-2147479015)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int4 square(int4 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x) &gt;= x</c>. For example, <c>square(4294967295u)</c>
+        /// will return <c>1u</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static uint square(uint x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x) &gt;= x</c>. For example, <c>square(new uint2(4294967295u))</c>
+        /// will return <c>new uint2(1u)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static uint2 square(uint2 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x) &gt;= x</c>. For example, <c>square(new uint3(4294967295u))</c>
+        /// will return <c>new uint3(1u)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static uint3 square(uint3 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
+        /// Computes the component-wise square (x * x) of the input argument x.
+        /// </summary>
+        /// <remarks>
+        /// Due to integer overflow, it's not always guaranteed that <c>square(x) &gt;= x</c>. For example, <c>square(new uint4(4294967295u))</c>
+        /// will return <c>new uint4(1u)</c>.
+        /// </remarks>
+        /// <param name="x">Value to square.</param>
+        /// <returns>Returns the square of the input.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static uint4 square(uint4 x)
+        {
+            return x * x;
+        }
+
+        /// <summary>
         /// Packs components with an enabled mask to the left.
         /// </summary>
         /// <remarks>
@@ -4242,6 +4450,51 @@ namespace Unity.Mathematics
             return h | (ux & ~msk) >> 16;
         }
 
+        /// <summary>
+        /// Generate an orthonormal basis given a single unit length normal vector.
+        /// </summary>
+        /// <remarks>
+        /// This implementation is from "Building an Orthonormal Basis, Revisited"
+        /// https://graphics.pixar.com/library/OrthonormalB/paper.pdf
+        /// </remarks>
+        /// <param name="normal">Unit length normal vector.</param>
+        /// <param name="basis1">Output unit length vector, orthogonal to normal vector.</param>
+        /// <param name="basis2">Output unit length vector, orthogonal to normal vector and basis1.</param>
+        public static void orthonormal_basis(float3 normal, out float3 basis1, out float3 basis2)
+        {
+            var sign = normal.z >= 0.0f ? 1.0f : -1.0f;
+            var a = -1.0f / (sign + normal.z);
+            var b = normal.x * normal.y * a;
+            basis1.x = 1.0f + sign * normal.x * normal.x * a;
+            basis1.y = sign * b;
+            basis1.z = -sign * normal.x;
+            basis2.x = b;
+            basis2.y = sign + normal.y * normal.y * a;
+            basis2.z = -normal.y;
+        }
+
+        /// <summary>
+        /// Generate an orthonormal basis given a single unit length normal vector.
+        /// </summary>
+        /// <remarks>
+        /// This implementation is from "Building an Orthonormal Basis, Revisited"
+        /// https://graphics.pixar.com/library/OrthonormalB/paper.pdf
+        /// </remarks>
+        /// <param name="normal">Unit length normal vector.</param>
+        /// <param name="basis1">Output unit length vector, orthogonal to normal vector.</param>
+        /// <param name="basis2">Output unit length vector, orthogonal to normal vector and basis1.</param>
+        public static void orthonormal_basis(double3 normal, out double3 basis1, out double3 basis2)
+        {
+            var sign = normal.z >= 0.0 ? 1.0 : -1.0;
+            var a = -1.0 / (sign + normal.z);
+            var b = normal.x * normal.y * a;
+            basis1.x = 1.0 + sign * normal.x * normal.x * a;
+            basis1.y = sign * b;
+            basis1.z = -sign * normal.x;
+            basis2.x = b;
+            basis2.y = sign + normal.y * normal.y * a;
+            basis2.z = -normal.y;
+        }
 
         /// <summary>Returns a uint hash from a block of memory using the xxhash32 algorithm. Can only be used in an unsafe context.</summary>
         /// <param name="pBuffer">A pointer to the beginning of the data.</param>