+++ src/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 * Added `[Il2CppEagerStaticClassConstruction]` to Unity.Mathematics types to run static constructors at startup. This improves IL2CPP performance slightly for types that have static constructors.
+* Added `math.square()` to compute the square (x * x).
+* Added `math.orthonormal_basis()` to compute an orthonormal basis from a single unit length vector.
 
 ### Changed
 * Changed noise documentation in comments to xmldoc comments.